Remove functions that operate with scalars directly (addc, mulc ecc...)
This update results in a much cleaner API where every binary operation
scalar or not is handled in the same way. In the C API scalars are
wrapped on the stack with alloca resulting in very little overhead.
Also, restrict has been removed from binary operations because they must
support assignment syntax like x += y (xa and out are the same tensor).