Add general purpose memory allocator (#20)
Introduced two new concepts:
- backend: a device with a list capabilities like allocating and freeing
memory
- allocator: allocate and free objects on a buffer created by a backend
For now DSC uses an hybrid approach with a bigger memory pool and a
general purpose allocator to create tensors with their own independent
lifetime and a smaller pool handled like an arena to store
intermediates.
Now #6 is required to validate the performance of this mechanism.