Binary ops in CUDA, drop FFT support for now
This project needs some real-world use-cases. For this reason I think
moving towards AI inference is a good move.
The goal is simple: develop a Python (first) API that lets you define a
neural network with a PyTorch-like syntax.
This project will become what YAMI was supposed to be from the start!