Load state_dict from url
* Add helpers to download safetensors files from url and create dsc
Tensors
* Parameters are now lazy-loaded from a state dict (i.e. when the model
is created only the Tensor descriptor is allocated, not the actual data
buffer. The buffer is then taken from the corresponding Tensor in the
state dict)
* dsc_ctx_free is no longer called when deleting the context object in
Python because it could be called before all the tensors have been freed
causing segfaults. This is a bug that I'll have to fix later on